Chore: move out get all tests to a separate file
